+++ work1/WuJunkai2004/.gitignore
@@ -1,3 +1,8 @@
 #mnist files
 */mnist/*.gz
 
+#all of datebase files
+*.db
+
+#debug file
+debug.*
