+++ work1/WuJunkai2004/3-native-bayes/data_clean.py
@@ -1,6 +1,7 @@
 import cv
 import dblite
 import read
+import param
 
 data = {}
 
@@ -8,24 +9,22 @@ db = dblite.SQL('clean.db')
 
 for num in range(10):
     try:
-        parts = ["part_{}".format(i) for i in range(1,7)]
-        db['num_{}'.format(num)].create( *parts )
+        parts = ["part_{}".format(i) for i in range(1, 1 + param.threshold.split.width * param.threshold.split.height )]
     except:
         pass
+
+    db['num_{}'.format(num)].create( *parts )
 
     for item in read.get_num_image(num):
         img = cv.image(item)
         img = img.clip_image(img.get_image_info())
         img = img.center_image()
 
-        density = [
-            img.get_region(0, 0).get_density(),
-            img.get_region(0, 1).get_density(),
-            img.get_region(1, 0).get_density(),
-            img.get_region(1, 1).get_density(),
-            img.get_region(2, 0).get_density(),
-            img.get_region(2, 1).get_density()
-        ]
+        density = []
+        for h in range(param.threshold.split.height):
+            for w in range(param.threshold.split.width):
+                density.append( img.get_region(h, w).get_density() )
+
 
         density = cv.get_absolute_density(density)
         density = cv.get_visualise_density(density)
@@ -37,3 +36,5 @@ for num in range(10):
                 data[item] += 1
 
 print(data)
+
+input()
