+++ work1/WuJunkai2004/3-native-bayes/requirements.txt
@@ -1,0 +1,1 @@
+numpy
